backend/agent-core/requirements.txt
fastapi
uvicorn
chromadb
google-genai
python-dotenv
